\name{example}
\alias{example}
\title{An example with n=102 samples and p=102 nodes}
\description{A simulated dataset according to 
a Gaussian linear mechanism specified by a DAG}
\format{
example is a list with three components (in the given order):
\describe{
\item{Y.n}{: the n by p data matrix}
\item{true.dir}{: the p by p adjacency matrix of the data generating dag}
\item{true.ske}{: the p by p adjacency matrix of the skeleton graph of the data generating dag}
 }
}

\usage{data(example)}
\keyword{datasets}